REFACTORIZED the components storage
