+++ src/components/Home/BlogCard.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { posts } from "../../../data/posts";
+import { posts } from "../../data/posts";
 
 export const BlogCard = () => {
   return (
