+++ src/components/Home/index.js
@@ -1,6 +1,6 @@
-export * from "./Benefits/Benefits";
-export * from "./BlogSection/BlogSection";
-export * from "./BlogCard/BlogCard";
-export * from "./Hero/Hero";
-export * from "./Subscription/Subscription";
-export * from "./WtppButton/WtppButton";
+export * from "./Benefits";
+export * from "./BlogSection";
+export * from "./BlogCard";
+export * from "./Hero";
+export * from "./Subscription";
+export * from "./WtppButton";
